feat(WIP): Adiciona melhorias - Adiciona estrutura inicial de CI comgithub actions - Adiciona sufire-report para geracao de relatorio de testes - Adiciona commons-io para remover pasta de forma recursiva - Muda estrutura do projeto
