+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Build
       run: mvn clean install -DskipTests
     - name: Unit test
-      run: mvn -B org.jacoco:jacoco-maven-plugin:0.8.11:prepare-agent test sufire-report:report org.jacoco:jacoco-maven-plugin:0.8.11:report
+      run: mvn -B org.jacoco:jacoco-maven-plugin:0.8.11:prepare-agent test org.apache.maven.plugins:maven-surefire-report-plugin:3.2.5:report org.jacoco:jacoco-maven-plugin:0.8.11:report
     - name: Quality Gate
       uses: SonarSource/sonarcloud-github-action@master
       env:
