+++ pom.xml
@@ -10,13 +10,13 @@
   </parent>
 
   <groupId>com.murilohenzo</groupId>
-  <artifactId>atom-pet-api</artifactId>
+  <artifactId>pet-api</artifactId>
   <version>0.0.1-SNAPSHOT</version>
-  <name>atom-pet-api</name>
+  <name>pet-api</name>
   <description>Pet API&lt;</description>
   <properties>
     <java.version>21</java.version>
-    <package>${project.groupId}.atom.petapi</package>
+    <package>${project.groupId}.petapi</package>
     <openapi.version>1.7.0</openapi.version>
     <org.mapstruct.version>1.5.5.Final</org.mapstruct.version>
   </properties>
@@ -113,6 +113,12 @@
       <artifactId>jackson-databind-nullable</artifactId>
       <version>0.2.6</version>
     </dependency>
+
+    <dependency>
+      <groupId>commons-io</groupId>
+      <artifactId>commons-io</artifactId>
+      <version>2.15.1</version>
+    </dependency>
   </dependencies>
 
   <build>
@@ -157,7 +163,6 @@
           </execution>
         </executions>
       </plugin>
-
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
@@ -181,7 +186,6 @@
           </annotationProcessorPaths>
         </configuration>
       </plugin>
-
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-dependency-plugin</artifactId>
@@ -204,6 +208,23 @@
                   <destFileName>${project.artifactId}.jar</destFileName>
                 </artifactItem>
               </artifactItems>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-surefire-report-plugin</artifactId>
+        <version>3.2.5</version>
+        <executions>
+          <execution>
+            <id>sufire-report</id>
+            <goals>
+              <goal>report</goal>
+            </goals>
+            <phase>test</phase>
+            <configuration>
+              <alwaysGenerateSurefireReport>false</alwaysGenerateSurefireReport>
             </configuration>
           </execution>
         </executions>
