--- src/main/java/com/murilohenzo/atom/petapi/domain/entities/Gender.java
@@ -1,6 +1,0 @@
-package com.murilohenzo.atom.petapi.domain.entities;
-
-public enum Gender {
-  FEMALE,
-  MALE
-}
