--- src/main/java/com/murilohenzo/atom/petapi/domain/entities/Status.java
@@ -1,5 +1,0 @@
-package com.murilohenzo.atom.petapi.domain.entities;
-
-public enum Status {
-  AVAILABLE, PENDING, SOLD
-}
