--- src/main/java/com/murilohenzo/atom/petapi/domain/service/PetPhotoRecord.java
@@ -1,6 +1,0 @@
-package com.murilohenzo.atom.petapi.domain.service;
-
-import org.springframework.http.MediaType;
-
-public record PetPhotoRecord(byte[] image, MediaType mediaType) {
-}
