--- src/main/java/com/murilohenzo/atom/petapi/domain/service/PetService.java
@@ -1,31 +1,0 @@
-package com.murilohenzo.atom.petapi.domain.service;
-
-import com.murilohenzo.atom.petapi.domain.entities.Pet;
-import com.murilohenzo.atom.petapi.domain.entities.Status;
-import com.murilohenzo.atom.petapi.domain.repository.PetRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
-
-@RequiredArgsConstructor
-public class PetService {
-
-  private final PetRepository repository;
-
-  @Transactional
-  public Pet create(Pet pet) {
-    return repository.save(pet);
-  }
-
-  @Transactional(readOnly = true)
-  public List<Pet> findPetsByStatus(Status status) {
-    return repository.findByStatus(status);
-  }
-
-  @Transactional(readOnly = true)
-  public Optional<Pet> findById(Long petId) {
-    return repository.findById(petId);
-  }
-}
