+++ src/main/java/com/murilohenzo/petapi/domain/entities/Gender.java
@@ -1,0 +1,6 @@
+package com.murilohenzo.petapi.domain.entities;
+
+public enum Gender {
+  FEMALE,
+  MALE
+}
