+++ src/main/java/com/murilohenzo/petapi/domain/entities/Status.java
@@ -1,0 +1,5 @@
+package com.murilohenzo.petapi.domain.entities;
+
+public enum Status {
+  AVAILABLE, PENDING, SOLD
+}
