+++ src/main/java/com/murilohenzo/petapi/domain/repository/StorageRepository.java
@@ -1,0 +1,11 @@
+package com.murilohenzo.petapi.domain.repository;
+
+import java.io.InputStream;
+
+public interface StorageRepository {
+  public String saveImage(String fileName, InputStream is);
+
+  public byte[] getImage(String imgReferenceId);
+
+  public boolean deleteImage(String imgReferenceId);
+}
