+++ src/main/java/com/murilohenzo/petapi/domain/service/PetPhotoRecord.java
@@ -1,0 +1,6 @@
+package com.murilohenzo.petapi.domain.service;
+
+import org.springframework.http.MediaType;
+
+public record PetPhotoRecord(byte[] image, MediaType mediaType) {
+}
